template/pyproject.toml
[project]
name = "adk-a2a-agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.20",
    "google-adk[a2a]>=1.20.0",
    "google-cloud-firestore>=2.19.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-requests>=0.58b0",
    "opentelemetry-instrumentation-google-genai>=0.4b0",
    "google-adk-redis>=0.1.2",
    "adk-redis[memory]>=0.0.3",
    "redisvl>=0.6.0",
    "asyncpg>=0.31.0",
    "motor>=3.7.1",
]
[dependency-groups]
develop = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "vulture>=2.14",
]
